+++ catalog/app/utils/BucketPreferences/index.ts
@@ -1,5 +1,6 @@
 export type {
   ActionPreferences,
+  AthenaPreferences,
   BucketPreferences,
   NavPreferences,
   SourceBuckets,